fix: get rid of middle class and code smell
