+++ .aws/task-definition-template.json
@@ -178,7 +178,7 @@
           "valueFrom": "application_cron_revalcurrentplacementjob-sync-${environment}"
         },
         {
-          "name": "APPLICATION_CRON_POSTFUNDINGSYNCJOB",
+          "name": "APPLICATION_CRON_POSTFUNDING_STATUS_SYNCJOB",
           "valueFrom": "application_cron_postfundingjob-sync-${environment}"
         },
         {
