+++ src/main/java/uk/nhs/tis/sync/api/JobResource.java
@@ -20,7 +20,7 @@ import uk.nhs.tis.sync.job.PersonPlacementEmployingBodyTrustJob;
 import uk.nhs.tis.sync.job.PersonPlacementTrainingBodyTrustJob;
 import uk.nhs.tis.sync.job.PersonRecordStatusJob;
 import uk.nhs.tis.sync.job.PostEmployingBodyTrustJob;
-import uk.nhs.tis.sync.job.PostFundingSyncJob;
+import uk.nhs.tis.sync.job.PostFundingStatusSyncJob;
 import uk.nhs.tis.sync.job.PostTrainingBodyTrustJob;
 import uk.nhs.tis.sync.job.RunnableJob;
 import uk.nhs.tis.sync.job.person.PersonElasticSearchSyncJob;
@@ -48,7 +48,7 @@ public class JobResource {
   private final PersonRecordStatusJob personRecordStatusJob;
   private RevalCurrentPmSyncJob revalCurrentPmSyncJob;
   private RevalCurrentPlacementSyncJob revalCurrentPlacementSyncJob;
-  private PostFundingSyncJob postFundingSyncJob;
+  private PostFundingStatusSyncJob postFundingStatusSyncJob;
 
   @Autowired
   private JobRunningListener jobRunningListener;
@@ -83,15 +83,15 @@ public class JobResource {
   }
 
   @Autowired
-  public void setPostFundingSyncJob(PostFundingSyncJob postFundingSyncJob) {
-    this.postFundingSyncJob = postFundingSyncJob;
+  public void setPostFundingStatusSyncJob(PostFundingStatusSyncJob postFundingStatusSyncJob) {
+    this.postFundingStatusSyncJob = postFundingStatusSyncJob;
   }
 
   /**
    * GET /jobs/status : Get all the status of 8 jobs.
    *
    * @return map of the status for most jobs. eg. {"personPlacementEmployingBodyTrustJob", "true"},
-   * which means personPlacementEmployingBodyTrustJob is currently running.
+   *    which means personPlacementEmployingBodyTrustJob is currently running.
    */
   @GetMapping("/jobs/status")
   @PreAuthorize("hasPermission('tis:sync::jobs:', 'View')")
@@ -112,7 +112,7 @@ public class JobResource {
     if (revalCurrentPlacementSyncJob != null) {
       statusMap.put("revalCurrentPlacementJob", revalCurrentPlacementSyncJob.isCurrentlyRunning());
     }
-    statusMap.put("postFundingSyncJob", postFundingSyncJob.isCurrentlyRunning());
+    statusMap.put("postFundingStatusSyncJob", postFundingStatusSyncJob.isCurrentlyRunning());
     return ResponseEntity.ok().body(statusMap);
   }
 
@@ -137,7 +137,7 @@ public class JobResource {
    *
    * @param name the name of the job to run
    * @return status of the requested job : "Already running" - the job has been running before
-   * triggering it "Just started" - the job has been started by this request
+   *    triggering it "Just started" - the job has been started by this request
    */
   @PutMapping("/job/{name}")
   @PreAuthorize("hasPermission('tis:sync::jobs:', 'Update')")
@@ -188,8 +188,8 @@ public class JobResource {
           return ResponseEntity.badRequest().body(JOB_NOT_FOUND);
         }
         break;
-      case "postFundingSyncJob":
-        status = ensureRunning(postFundingSyncJob, params);
+      case "postFundingStatusSyncJob":
+        status = ensureRunning(postFundingStatusSyncJob, params);
         break;
       default:
         return ResponseEntity.badRequest().body(JOB_NOT_FOUND);
