+++ src/main/resources/static/index.html
@@ -45,9 +45,9 @@
     <p id="personRecordStatusJob_status" class="box-3"></p>
   </div>
   <div class="container-8">
-    <h2 class="box-1">Post Funding Sync Job</h2>
-    <button type="button" id="postFundingSyncJob" class="box-2">Run job</button>
-    <p id="postFundingSyncJob_status" class="box-3"></p>
+    <h2 class="box-1">Post Funding Status Sync Job</h2>
+    <button type="button" id="postFundingStatusSyncJob" class="box-2">Run job</button>
+    <p id="postFundingStatusSyncJob_status" class="box-3"></p>
   </div>
 </div>
 <br>
